extract client to its own file
