+++ src/goose/toolkit/language_servers/sync.py
@@ -1,10 +1,10 @@
-from goose.language_server.base import SyncLanguageServer
-from goose.language_server.config import Language, MultilspyConfig
+from goose.language_server.client import LanguageServerClient
+from goose.language_server.config import MultilspyConfig
 from goose.language_server.implementations.jedi import JediServer
 from goose.language_server.logger import MultilspyLogger
 
 
-sls = SyncLanguageServer()
+sls = LanguageServerClient()
 sls.register_language_server(JediServer.from_env(config=MultilspyConfig(), logger=MultilspyLogger()))
 with sls.start_servers() as _:
     d = sls.request_definition("/Users/lalvoeiro/Development/goose/src/goose/build.py", 41, 28)
